refactor: consolidate upload logic
