+++ server/routes.ts
@@ -33,7 +33,7 @@ import { cleanupOrphanedProducts } from "./cleanup";
 import { success, failure } from "./lib/responses";
 import { TICKET_CENT_VALUE } from "../config/business";
 
-import { registerProfileImageRoutes } from "./lib/simple-profile-upload";
+import { registerProfileImageRoutes } from "./lib/profile-upload";
 
 function extractAsin(url: string): string {
   const asinPattern = /(?:\/dp\/|\/gp\/product\/|\/ASIN\/|%2Fdp%2F)([A-Z0-9]{10})/i;
